Complete refactoring component organization and directory structure
