+++ client/components/ItemEntry.jsx
@@ -5,10 +5,9 @@ import { Link } from 'react-router';
 import $ from 'jquery';
 import prettyDate from 'dateformat';
 import { checkAuthentication } from './../actions';
+import CountdownTimer from './CountDownTimer.jsx';
 
-import CountdownTimer from './CountdownTimer.jsx';
-
-export default class ItemEntry extends React.Component {
+export default class PopularItemEnry extends React.Component {
 
   componentDidMount() {
     this.props.getUser();
@@ -95,4 +94,4 @@ var mapDispatchToProps = function(dispatch) {
   }
 };
 
-module.exports = connect(mapStateToProps, mapDispatchToProps)(ItemEntry);
+module.exports = connect(mapStateToProps, mapDispatchToProps)(PopularItemEnry);
