+++ client/components/personal-page/ListedItems.jsx
@@ -2,8 +2,8 @@ import React, { Component, PropTypes } from 'react';
 import ReactDOM from 'react-dom';
 import { connect } from 'react-redux';
 import $ from 'jquery';
-import { checkAuthentication } from './../actions';
-import ItemEntry from './ItemEntry.jsx';
+import { checkAuthentication } from './../../actions';
+import ItemEntry from './../ItemEntry.jsx';
 
 class ListedItems extends Component {
   static defaultProps = {
